Move API into ApiModel.
In reality it is part of the ApiModel and the ApiDump namespace was not
adequate. Dumping is part of the operation in the API.

If needed we could re-add it with different kinds of "Dumpers"
