+++ bin/Dumper.cs
@@ -12,7 +12,7 @@ public class Dumper
         }
         Exml.Logging.Logger.AddConsoleLogger();
 
-        var api = Exml.ApiDump.API.Parse(args[0]);
+        var api = Exml.ApiModel.API.Parse(args[0]);
 
         var filename = Path.GetFileName(args[0]);
         filename += ".api";
@@ -24,7 +24,7 @@ public class Dumper
         Dump("api.yaml", api);
     }
 
-    public static void Dump(string filename, Exml.ApiDump.API api)
+    public static void Dump(string filename, Exml.ApiModel.API api)
     {
         using (System.IO.StreamWriter file = new System.IO.StreamWriter(filename))
         {
